implement default behavior for the cards in the state trait
